Use open-ended version for supported Oscar version
Theoretically, it's nicer to limit the range of supported Oscar
versions. But in practice it becomes hard to prepare an extension for a
unreleased Oscar version if the range is limited. That's because
easy_install/pip will downgrade the used Oscar version again.

Long story short, experience shows an open-ended approach is best.