Send more identifiers to the payment platform
When working with postback notifications, we cannot rely on the
existence of the Oscar `CheckoutSession` to fetch the current order
when coming back from Adyen. Our solution is to pass all the relevant
identifiers (`customer_id`, `basket_id` and `order_number`) into the
`merchantReference` Adyen field, which is systematically sent back
to the application, whether using a GET or a POST method.