Add config class to get settings
We will need to be able to alter settings based on e.g. the request, so
as a first step, I'm introducing a layer between Django settings and the
Adyen code.

We introduce the concept of a config class, that can be set from the
Django settings. It defaults to a supplied config class that just
fetches the needed values from the Django settings as before. But now we
have the possibility of switiching out the config class to fetch those
values from elsewhere.