Pass request into config class
The idea is that we want to change certain configuration values based on
the request. To get this working, we need to change the interface for
the config class to pass the request to most of it's methods.

To get the request object, we need to also alter the facade's and
scaffold's interface to more often include the request.

For this commit, I had a look at how data gets passed from e.g. Oscaro
to the scaffold. Unfortunately I discovered an anti-pattern, where a lot
of required data is passed in via an undocumented order_data dictionary.

That needs to be cleaned up separately.