Remove Math from authors
The extension has had multiple people working on it now, and Oscaro is
the organization ultimately guiding it. I updated the package
information to reflect that.

I also removed Oscar as a dependecy. It's quite obvious (as is Django).
From experience, it also makes a project's installing requirements phase
more predictable.