Safeguard against _compute_hash returning None
I was bothered by params.get(self.HASH_FIELD) because I'd expect the
hash field to always be present. But I realized it's better to allow it
not being present, but loudly raising an InvalidTransactionException,
irrespective whehter self.hash() returns anything or not (which it might
not do accidentally).

This is probably minor nitpicking, but better be safe than sorry!