Handle ERROR notifications from Adyen
The point of this PR really is to address a bug with ERROR notifications
from Adyen. Oscaro employees can check Sentry:
http://sentry-prod.oscaro.com/oshop/oshop-prod/group/1310/events/3457809/

The bug is two-fold. Error messages were forgotten in the status map.
Note that PAYMENT_RESULT_ERROR is already present in the codebase.

The other issue was that the lookup in ADYEN_TO_COMMON_STATUSES
was implemented to be forgiving (using get()). Hence the
traceback above is somewhat subtle; it lead to status being returned
as None.