Use Oscar dynamic loading with get_class
One of the main concern about django-oscar-adyen is that one may want
to override parts of the application. For example, how the merchantData
are used, or what to do to record a transaction.

Since there is no one-size-fit-all, I reuse here a very convenient tool
from django-oscar: `oscar.core.loading.get_class`. This tool allows us
to provide a base interface, that can be used almost out of the box,
and make possible anyone to override parts of it.

This is only the first step to a real modular django-oscar plugin, but
that's a good first step.