Split get_form_fields in 2 parts
The first part is handled by `get_field_specs`: this method is
responsible for returning a dict which contains the expected keys.

The second part is handled by the Facade, with its method
`build_payment_form_fields` which take the request and the reslt from
`get_field_specs`.

The `get_field_specs` is also split into multiple methods:

* `get_field_merchant_return_data` handles the content of the field
  `merchantReturnData` (by default it's the order's `amount`)
* `get_field_return_url` handles the content of the field `resURL`.

These changes are backward compatible.