Add adyen.signers module
This module contains:

* An AbstractSigner class to define a signer interface,
* A signer implementation using the SHA-1 algorithm.

Since `django-oscar-adyen` supports only the SHA-1 algorithm at the
moment, this first commit only contains this version. A future commit
will contain new algorithm, once the `adyen.signers` module is used.

The unit tests take data from other tests to ensure that the HMACSha1
signer output the same result as the current implementation, provided
with the same input.

The documentation only contains the auto-doc part, and a new chapter to
describe how to configure and to use signers should be written later.