Use `adyen.signers` into the gateway
The `adyen.signers` contains signers (well, only one at the moment).
The configuration class used to build the `adyen.gateway.Gateway`
object instantiate one of these signers, and give it to the gateway.

The new setting is `ADYEN_SIGNER_BACKEND`: it is the python path to the
signer class. By default and to be backward compatible it uses the
`adyen.signers.HMACSha1` signer.

Then, when the gateway instantiate a `PaymentFormRequest` object, the
signature `merchantSig` is computed by the gateway's signer. When a
`PaymentRedirection` is validated (using `PaymentRedirection.validate`)
the gateway's signer is used again.

This is backward compatible with 0.6.0 and should not be an issue for
plugin users.

The unit tests have been modified to ensure that we keep this version
backward compatible with 0.6.0:

* Settings use a default value,
* Same signatures used in tests as before,
* New tests to ensure how the settings class behave.

The documentation contains the description of the new setting
`ADYEN_SIGNER_BACKEND`, with its default value.