Signer implementation using the SHA-256 algorithm.
The SHA-256 algorithm is used by Adyen with a very simple mechanism:
all the keys used must be signed, and key are ordered by their name.

This implementation has not been tested in production (unlike the
SHA-1 algorithm), but the tests make sure to use sample data from the
Adyen documentation.